Refactor: 명칭 변경
